[FIX] website_sale_wishlist: fix "save for later" button in cart
Whenever the cart changes, the cart lines are re-rendered. As a result, the
interaction handling the cart lines needs to be restarted at each re-render.
However, the "save for later" button is handled in a separate interaction,
which wasn't restarted at each re-render (because its selector was too broad).
Consequently, the "save for later" button only worked before the first
re-render.

This PR splits the interaction handling the "save for later" button to make sure
it's restarted whenever the cart lines are re-rendered. As an added benefit, the
scope of the new interactions is now more specific (which is a best practice).
